infrastructure/testdriver/actions.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>TestDriver actions method</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<input id="checkbox" type="checkbox">

<script>
async_test(function(t) {
  var target = document.getElementById("checkbox");
  target.addEventListener("pointermove", t.step_func_done(function() {
  }));
  assert_true(test_driver.apple === "10");

  test_driver.actions({}).then((function() {
//   test_driver.actions([{"type": "click", "args": [{"type": "element", "arg": target}]}, {"type": "perform"}]).then((function() {
      return;
  })).catch(t.unreached_func("actions failed"));
});
</script>
